move api request to appropriate file
