+++ astralAideKillbot/embed_creator.py
@@ -1,8 +1,7 @@
 import discord
-from collections import Counter
 from .api_request import *
 from .config import TARGET_DISCORD_CHANNEL_ID, TARGET_ENTITY_ID, TARGET_ENTITY
-from .utils import format_currency, capitalize_and_replace, log_time, get_current_time
+from .utils import * 
 
 semaphore = asyncio.Semaphore(5)  # Limit concurrent tasks
 
@@ -90,36 +89,3 @@ async def send_killmail_embed(bot, killmail):
         
         log_time(start_time, killmail.get('killmail_id'))
 
-async def organize_killmail_data(killmail):
-    esi_url = f"https://esi.evetech.net/latest/killmails/{killmail.get('killID')}/{killmail.get('hash')}/"
-    esi_data = await fetch_data(esi_url)
-
-    zkill_url = f"https://zkillboard.com/api/killID/{killmail.get('killID')}/"
-    zkill_data = await fetch_data(zkill_url)
-
-    if zkill_data and 'zkb' in zkill_data[0]:
-        esi_data['zkb'] = zkill_data[0]['zkb']
-
-    return esi_data
-
-async def get_top_kills_losses(character_id):
-    kills, losses = await get_kill_loss_data(character_id)
-    kill_ship_types = [await fetch_killmail_ship_type(kill) for kill in kills]
-    loss_ship_types = [await fetch_killmail_ship_type(loss) for loss in losses]
-    
-    top_3_kill_ship_types = Counter(kill_ship_types).most_common(3)
-    top_3_loss_ship_types = Counter(loss_ship_types).most_common(3)
-
-    top_3_kill_names = [[f"{await get_ship_type_name(ship_type) or 'N/A'} ({count})", ship_type] for ship_type, count in top_3_kill_ship_types]
-    top_3_loss_names = [[f"{await get_ship_type_name(ship_type) or 'N/A'} ({count})", ship_type] for ship_type, count in top_3_loss_ship_types]
-    
-    return top_3_kill_names, top_3_loss_names
-
-async def fetch_killmail_ship_type(kill):
-    try:
-        killmail_id = kill['killmail_id']
-        killmail_hash = kill['zkb']['hash']
-        killmail_details = await fetch_killmail_details(killmail_id, killmail_hash)
-        return killmail_details['victim']['ship_type_id']
-    except (KeyError, TypeError):
-        return None
